Developer

+++ src/routes/movies/delete.js
@@ -10,8 +10,14 @@ const app = express();
 app.use(express.json());
 
 
-app.delete('/', async (req, res) => {
+app.delete('/delete/:id', async (req, res) => {
     try {
+        // Verifique se o filme existe antes de tentar excluí-lo
+        const movieToDelete = await Movie.findById(req.params.id);
+        if (!movieToDelete) {
+            return res.status(404).json({ error: 'Filme não encontrado' });
+        }
+
         const deleteMovie = await Movie.findByIdAndDelete(req.params.id);
         res.json(deleteMovie);
         res.json({ message: 'Filme deletado com sucesso' });
@@ -19,5 +25,6 @@ app.delete('/', async (req, res) => {
         res.status(500).json({ error: 'Erro ao deletar filme' });
     }
 });
+
 
 export default app;
